pyproject.toml
[project]
name = "nlp100"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gensim>=4.4.0",
    "ginza>=5.2.0",
    "google-generativeai>=0.8.5",
    "ipadic>=1.0.0",
    "ipykernel>=7.1.0",
    "ja-ginza>=5.2.0",
    "janome>=0.5.0",
    "matplotlib>=3.10.7",
    "mecab-python3>=1.0.12",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sudachidict-core>=20251022",
    "sudachipy>=0.6.10",
]
